UNG-926 Adding keystore to cassandra conn
