+++ src/main/resources/application-docker.conf
@@ -20,6 +20,11 @@ acctSystem {
       consistencyLevel = ${ACCT_SVC_CS_CONSISTENCY_LEVEL}
       serialConsistencyLevel = ${ACCT_SVC_CS_SERIAL_CONSISTENCY_LEVEL}
       withSSL: ${ACCT_SVC_CS_WITH_SSL}
+      //When true, truststore and truststore password must be provided
+      trustStore: ""
+      trustStore: ${?ACCT_SVC_CS_TRUST_STORE}
+      trustStorePassword: ""
+      trustStorePassword: ${?ACCT_SVC_CS_TRUST_STORE_PASSWORD}
       username: ${ACCT_SVC_CS_USERNAME}
       password: ${ACCT_SVC_CS_PASSWORD}
