+++ src/main/resources/application.base.conf
@@ -20,6 +20,9 @@ acctSystem {
       consistencyLevel = "LOCAL_ONE"
       serialConsistencyLevel = "SERIAL"
       withSSL: false
+      //When true, truststore and truststore password must be provided
+      trustStore: ""
+      trustStorePassword: ""
       username: ""
       password: ""
