+++ src/main/scala/com/ubirch/services/cluster/ClusterService.scala
@@ -2,6 +2,8 @@ package com.ubirch
 package services.cluster
 
 import java.net.InetSocketAddress
+import java.nio.file.{ Files, Paths }
+import java.security.KeyStore
 
 import com.datastax.driver.core._
 import com.datastax.driver.core.policies.RoundRobinPolicy
@@ -9,6 +11,7 @@ import com.typesafe.config.Config
 import com.ubirch.ConfPaths.CassandraClusterConfPaths
 import com.ubirch.util.URLsHelper
 import javax.inject._
+import javax.net.ssl.{ SSLContext, TrustManagerFactory }
 
 /**
   * Component that contains configuration-related values.
@@ -46,6 +49,23 @@ trait ClusterService extends ClusterConfigs {
   val poolingOptions: PoolingOptions
   val queryOptions: QueryOptions
   val cluster: Cluster
+
+  def buildSSLOptions(trustStorePath: String, trustStorePassword: String): RemoteEndpointAwareJdkSSLOptions = {
+    val trustStore = KeyStore.getInstance("JKS")
+    closableTry(Files.newInputStream(Paths.get(trustStorePath)))(_.close()) { stream =>
+      trustStore.load(stream, trustStorePassword.toCharArray)
+    }
+
+    val trustManagerFactory = TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm)
+    trustManagerFactory.init(trustStore)
+
+    val sslContext = SSLContext.getInstance("TLS")
+    sslContext.init(null, trustManagerFactory.getTrustManagers, null)
+
+    RemoteEndpointAwareJdkSSLOptions.builder().withSSLContext(sslContext).build()
+
+  }
+
 }
 
 /**
@@ -60,10 +80,12 @@ class DefaultClusterService @Inject() (config: Config) extends ClusterService wi
   val maybeConsistencyLevel: Option[ConsistencyLevel] = checkConsistencyLevel(config.getString(CONSISTENCY_LEVEL))
   val maybeSerialConsistencyLevel: Option[ConsistencyLevel] = checkConsistencyLevel(config.getString(SERIAL_CONSISTENCY_LEVEL))
   val withSSL: Boolean = config.getBoolean(WITH_SSL)
+  lazy val trustStorePath: String = config.getString(TRUST_STORE)
+  lazy val trustStorePassword: String = config.getString(TRUST_STORE_PASSWORD)
   val username: String = config.getString(USERNAME)
   val password: String = config.getString(PASSWORD)
 
-  val poolingOptions = new PoolingOptions().setMaxQueueSize(1024)
+  val poolingOptions: PoolingOptions = new PoolingOptions().setMaxQueueSize(1024)
     .setMaxRequestsPerConnection(HostDistance.LOCAL, 32768)
     .setMaxRequestsPerConnection(HostDistance.REMOTE, 2000)
 
@@ -89,7 +111,7 @@ class DefaultClusterService @Inject() (config: Config) extends ClusterService wi
       .withClusterName("event-log")
 
     if (withSSL) {
-      builder.withSSL()
+      builder.withSSL(buildSSLOptions(trustStorePath, trustStorePassword))
     }
 
     builder.build()
